Commit Process

+++ SCL.py
@@ -134,8 +134,8 @@ def compute_class_connectiveity(pred_conn, label_conn, pred_num_conn,
 
     pred_conn = torch.tensor(pred_conn)
     label_conn = torch.tensor(label_conn)
-    pred_conn = F.one_hot(pred_conn, origin_pred_num_conn)
-    label_conn = F.one_hot(label_conn, label_num_conn)
+    pred_conn = F.one_hot(pred_conn.long(), origin_pred_num_conn)
+    label_conn = F.one_hot(label_conn.long(), label_num_conn)
 
     ious = torch.zeros((real_label_num, real_pred_num))
     pair_conn_sum = torch.tensor([0.], requires_grad=False)
